build: upgrade to Kotlin 2.3.21 / Compose 1.11.1 — flat 0.1.0 installer version
Compose Multiplatform 1.11 lifted the Gradle plugin's macOS "major > 0"
package-version check (JetBrains #5504), so the installer version is now a
single 0.1.0 across MSI/DEB/DMG and the per-OS override in :ui is gone —
the .dmg metadata finally matches the CLI's own version, no divergence.

Toolchain: Kotlin 2.1.0 -> 2.3.21, Compose 1.7.3 -> 1.11.1,
kotlinx-serialization 1.7.3 -> 1.11.0. Also drops a now-redundant UInt
conversion the 2.3 Win32 bindings flag in Lock.mingw.kt.

+++ core/src/mingwX64Main/kotlin/io/taskkling/core/Lock.mingw.kt
@@ -25,7 +25,7 @@ import platform.windows.UnlockFileEx
 internal actual fun <T> withSystemFileLock(lockPath: String, timeoutSeconds: Int, block: () -> T): T {
     val handle = CreateFileW(
         lockPath,
-        (GENERIC_READ.toUInt() or GENERIC_WRITE.toUInt()),
+        (GENERIC_READ or GENERIC_WRITE.toUInt()),
         (FILE_SHARE_READ.toUInt() or FILE_SHARE_WRITE.toUInt()),
         null,
         OPEN_ALWAYS.toUInt(),
